requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
openai
python-dotenv
numpy